refactor: moved settings to dataHandlers
