+++ src/PlotHandler/PlotHandler.hpp
@@ -1,35 +1,45 @@
-#ifndef _PLOTHANDLER_HPP
-#define _PLOTHANDLER_HPP
+#pragma once
 
+#include <chrono>
 #include <map>
-#include <memory>
-
-#include <string>
-
+#include <mutex>
+#include <thread>
 
 #include "Plot.hpp"
-#include "PlotHandlerBase.hpp"
+#include "ScrollingBuffer.hpp"
 
-class PlotHandler : public PlotHandlerBase
+class PlotHandler
 {
    public:
-	typedef struct Settings
+	std::shared_ptr<Plot> addPlot(const std::string& name);
+	bool removePlot(const std::string& name);
+	bool renamePlot(const std::string& oldName, const std::string& newName);
+	bool removeAllPlots();
+	std::shared_ptr<Plot> getPlot(std::string name);
+	bool eraseAllPlotData();
+	uint32_t getVisiblePlotsCount() const;
+	uint32_t getPlotsCount() const;
+	bool checkIfPlotExists(const std::string& name) const;
+	void setMaxPoints(uint32_t maxPoints);
+
+	class iterator
 	{
-		uint32_t sampleFrequencyHz = 100;
-		uint32_t maxPoints = 10000;
-		uint32_t maxViewportPoints = 5000;
-		bool refreshAddressesOnElfChange = false;
-		bool stopAcqusitionOnElfChange = false;
-		bool shouldLog = false;
-		std::string logFilePath = "";
-		std::string gdbCommand = "gdb";
-	} Settings;
+	   public:
+		using iterator_category = std::forward_iterator_tag;
+		explicit iterator(std::map<std::string, std::shared_ptr<Plot>>::iterator iter);
+		iterator& operator++();
+		iterator operator++(int);
+		bool operator==(const iterator& other) const;
+		bool operator!=(const iterator& other) const;
+		std::shared_ptr<Plot> operator*();
 
-	Settings getSettings() const;
-	void setSettings(const Settings& newSettings);
+	   private:
+		std::map<std::string, std::shared_ptr<Plot>>::iterator m_iter;
+	};
 
-   private:
-	Settings settings{};
+	iterator begin();
+	iterator end();
+
+   protected:
+	std::map<std::string, std::shared_ptr<Plot>> plotsMap;
 };
-
-#endif
